interface: Improve rendering of externals borders
Fix the rendering so that half of the external border lines are not
chopped off due to the fact that the lines are drawn centred on the
coordinates given to Cairo. This is possible by considering the whole
rendering area as needing to be BORDER_LEFT (the width) wider than it
actually is, and ensuring that all the cell size calculations are done
as floating point so as to not introduce rounding errors.

https://bugzilla.gnome.org/show_bug.cgi?id=732776

+++ src/main.h
@@ -84,8 +84,8 @@ typedef struct {
 	gdouble drawing_area_width;
 	gdouble drawing_area_height;
 
-	guint drawing_area_x_offset;
-	guint drawing_area_y_offset;
+	gdouble drawing_area_x_offset;
+	gdouble drawing_area_y_offset;
 
 	PangoFontDescription *normal_font_desc;
 	PangoFontDescription *painted_font_desc;
